# standart-server/openapi/bundles/v1.openapi.bundle.yaml
openapi: 3.0.3
info:
  title: Task Service API
  version: 1.0.0
  description: API для работы с данными о задачах!
servers:
  - url: /api/v1
    description: v1 prefix
tags:
  - name: Task
    description: Операции с данными о задачах — получение и обновление информации
paths:
  /tasks/{id}:
    parameters:
      - $ref: '#/components/parameters/id'
    get:
      summary: Get task by id
      operationId: GetTaskByID
      tags:
        - Task
      responses:
        '200':
          description: Task information successfully retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/task'
        '400':
          description: Bad request - validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/validation_error'
        '404':
          description: Task data for specified id not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'
    patch:
      summary: Patch task by id
      operationId: PatchTaskByID
      tags:
        - Task
      requestBody:
        description: Patch task with request body object
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/update_task'
        required: true
      responses:
        '200':
          description: Task successfully patched
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ok'
        '400':
          description: Bad request - validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/validation_error'
        '404':
          description: Task data for specified id not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'
components:
  parameters:
    id:
      name: id
      in: path
      required: true
      description: id задачи в формате uuid
      schema:
        type: string
        format: uuid
        example: acde070d-8c4c-4f0d-9d8a-162843c10333
  schemas:
    error:
      type: object
      required:
        - message
      properties:
        message:
          type: string
          description: Описание ошибки
          example: Ресурс не найден
    status:
      type: string
      description: |
        Текущее состояние жизненного цикла задачи: * `backlog` - Задача создана, но еще не взята в работу. * `in_progress` - Над задачей ведется работа. * `on_hold` - Работа приостановлена (ждем ответа, заблокирована). * `completed` - Задача успешно выполнена. * `canceled` - Задача отменена и не будет выполнена.
      enum:
        - backlog
        - in_progress
        - on_hold
        - completed
        - canceled
      example: in_progress
    task:
      type: object
      required:
        - id
        - title
        - is_active
        - status
        - created_at
      properties:
        id:
          type: string
          description: ID задачи
          example: acde070d-8c4c-4f0d-9d8a-162843c10333
          format: uuid
        title:
          type: string
          description: Заголовок задачи
          example: Задача 1
        is_active:
          type: boolean
          description: Отвечает за скрытость или видимость задачи
          example: true
        status:
          $ref: '#/components/schemas/status'
        created_at:
          type: string
          format: date-time
          description: Время создания задачи
          example: '2023-05-15T10:30:00Z'
    validation_error_item:
      type: object
      required:
        - field
        - error
      properties:
        field:
          type: string
          description: Поле у которого есть ошибка
          example: Username
        error:
          type: string
          description: Описание ошибки
          example: Username должен быть более 4 символов
    validation_error:
      type: object
      required:
        - message
      properties:
        message:
          type: string
          description: Описание ошибки
          example: Ошибка валидации
        errors:
          type: array
          description: Список полей с ошибками валидации
          items:
            $ref: '#/components/schemas/validation_error_item'
    update_task:
      type: object
      required:
        - email
      properties:
        title:
          type: string
          description: Заголовок задачи
          example: Задача 1
          nullable: true
          format: email
        email:
          type: string
          description: Адрес эл.почти
          example: test@gmail.com
          format: email
        is_active:
          type: boolean
          description: Отвечает за скрытость или видимость задачи
          example: true
          nullable: true
        status:
          $ref: '#/components/schemas/status'
          nullable: true
        ip:
          type: string
          description: IP адрес
          example: 127.0.0.1
          nullable: true
          format: ipv4
    ok:
      type: object
      required:
        - ok
      properties:
        ok:
          type: boolean
          description: Флаг - все хорошо или нет
          example: true