# standart-server/pkg/openapi/api/v1/api.gen.go
// Package api_v1 provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/oapi-codegen/oapi-codegen/v2 version v2.6.0 DO NOT EDIT.
package api_v1

import (
	"fmt"
	"net/http"
	"time"

	"github.com/go-chi/chi/v5"
	"github.com/oapi-codegen/nullable"
	"github.com/oapi-codegen/runtime"
	openapi_types "github.com/oapi-codegen/runtime/types"
)

// Defines values for Status.
const (
	Backlog    Status = "backlog"
	Canceled   Status = "canceled"
	Completed  Status = "completed"
	InProgress Status = "in_progress"
	OnHold     Status = "on_hold"
)

// Valid indicates whether the value is a known member of the Status enum.
func (e Status) Valid() bool {
	switch e {
	case Backlog:
		return true
	case Canceled:
		return true
	case Completed:
		return true
	case InProgress:
		return true
	case OnHold:
		return true
	default:
		return false
	}
}

// Error defines model for error.
type Error struct {
	// Message Описание ошибки
	Message string `json:"message"`
}

// Ok defines model for ok.
type Ok struct {
	// Ok Флаг - все хорошо или нет
	Ok bool `json:"ok"`
}

// Status Текущее состояние жизненного цикла задачи: * `backlog` - Задача создана, но еще не взята в работу. * `in_progress` - Над задачей ведется работа. * `on_hold` - Работа приостановлена (ждем ответа, заблокирована). * `completed` - Задача успешно выполнена. * `canceled` - Задача отменена и не будет выполнена.
type Status string

// Task defines model for task.
type Task struct {
	// CreatedAt Время создания задачи
	CreatedAt time.Time `json:"created_at"`

	// Id ID задачи
	Id openapi_types.UUID `json:"id"`

	// IsActive Отвечает за скрытость или видимость задачи
	IsActive bool `json:"is_active"`

	// Status Текущее состояние жизненного цикла задачи: * `backlog` - Задача создана, но еще не взята в работу. * `in_progress` - Над задачей ведется работа. * `on_hold` - Работа приостановлена (ждем ответа, заблокирована). * `completed` - Задача успешно выполнена. * `canceled` - Задача отменена и не будет выполнена.
	Status Status `json:"status"`

	// Title Заголовок задачи
	Title string `json:"title"`
}

// UpdateTask defines model for update_task.
type UpdateTask struct {
	// Email Адрес эл.почти
	Email openapi_types.Email `json:"email"`

	// Ip IP адрес
	Ip nullable.Nullable[string] `json:"ip,omitempty"`

	// IsActive Отвечает за скрытость или видимость задачи
	IsActive nullable.Nullable[bool] `json:"is_active,omitempty"`

	// Status Текущее состояние жизненного цикла задачи: * `backlog` - Задача создана, но еще не взята в работу. * `in_progress` - Над задачей ведется работа. * `on_hold` - Работа приостановлена (ждем ответа, заблокирована). * `completed` - Задача успешно выполнена. * `canceled` - Задача отменена и не будет выполнена.
	Status *Status `json:"status,omitempty"`

	// Title Заголовок задачи
	Title nullable.Nullable[openapi_types.Email] `json:"title,omitempty"`
}

// ValidationError defines model for validation_error.
type ValidationError struct {
	// Errors Список полей с ошибками валидации
	Errors *[]ValidationErrorItem `json:"errors,omitempty"`

	// Message Описание ошибки
	Message string `json:"message"`
}

// ValidationErrorItem defines model for validation_error_item.
type ValidationErrorItem struct {
	// Error Описание ошибки
	Error string `json:"error"`

	// Field Поле у которого есть ошибка
	Field string `json:"field"`
}

// Id defines model for id.
type Id = openapi_types.UUID

// PatchTaskByIDJSONRequestBody defines body for PatchTaskByID for application/json ContentType.
type PatchTaskByIDJSONRequestBody = UpdateTask

// ServerInterface represents all server handlers.
type ServerInterface interface {
	// Get task by id
	// (GET /tasks/{id})
	GetTaskByID(w http.ResponseWriter, r *http.Request, id Id)
	// Patch task by id
	// (PATCH /tasks/{id})
	PatchTaskByID(w http.ResponseWriter, r *http.Request, id Id)
}

// Unimplemented server implementation that returns http.StatusNotImplemented for each endpoint.

type Unimplemented struct{}

// Get task by id
// (GET /tasks/{id})
func (_ Unimplemented) GetTaskByID(w http.ResponseWriter, r *http.Request, id Id) {
	w.WriteHeader(http.StatusNotImplemented)
}

// Patch task by id
// (PATCH /tasks/{id})
func (_ Unimplemented) PatchTaskByID(w http.ResponseWriter, r *http.Request, id Id) {
	w.WriteHeader(http.StatusNotImplemented)
}

// ServerInterfaceWrapper converts contexts to parameters.
type ServerInterfaceWrapper struct {
	Handler            ServerInterface
	HandlerMiddlewares []MiddlewareFunc
	ErrorHandlerFunc   func(w http.ResponseWriter, r *http.Request, err error)
}

type MiddlewareFunc func(http.Handler) http.Handler

// GetTaskByID operation middleware
func (siw *ServerInterfaceWrapper) GetTaskByID(w http.ResponseWriter, r *http.Request) {

	var err error

	// ------------- Path parameter "id" -------------
	var id Id

	err = runtime.BindStyledParameterWithOptions("simple", "id", chi.URLParam(r, "id"), &id, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true, Type: "string", Format: "uuid"})
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "id", Err: err})
		return
	}

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.GetTaskByID(w, r, id)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// PatchTaskByID operation middleware
func (siw *ServerInterfaceWrapper) PatchTaskByID(w http.ResponseWriter, r *http.Request) {

	var err error

	// ------------- Path parameter "id" -------------
	var id Id

	err = runtime.BindStyledParameterWithOptions("simple", "id", chi.URLParam(r, "id"), &id, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true, Type: "string", Format: "uuid"})
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "id", Err: err})
		return
	}

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.PatchTaskByID(w, r, id)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

type UnescapedCookieParamError struct {
	ParamName string
	Err       error
}

func (e *UnescapedCookieParamError) Error() string {
	return fmt.Sprintf("error unescaping cookie parameter '%s'", e.ParamName)
}

func (e *UnescapedCookieParamError) Unwrap() error {
	return e.Err
}

type UnmarshalingParamError struct {
	ParamName string
	Err       error
}

func (e *UnmarshalingParamError) Error() string {
	return fmt.Sprintf("Error unmarshaling parameter %s as JSON: %s", e.ParamName, e.Err.Error())
}

func (e *UnmarshalingParamError) Unwrap() error {
	return e.Err
}

type RequiredParamError struct {
	ParamName string
}

func (e *RequiredParamError) Error() string {
	return fmt.Sprintf("Query argument %s is required, but not found", e.ParamName)
}

type RequiredHeaderError struct {
	ParamName string
	Err       error
}

func (e *RequiredHeaderError) Error() string {
	return fmt.Sprintf("Header parameter %s is required, but not found", e.ParamName)
}

func (e *RequiredHeaderError) Unwrap() error {
	return e.Err
}

type InvalidParamFormatError struct {
	ParamName string
	Err       error
}

func (e *InvalidParamFormatError) Error() string {
	return fmt.Sprintf("Invalid format for parameter %s: %s", e.ParamName, e.Err.Error())
}

func (e *InvalidParamFormatError) Unwrap() error {
	return e.Err
}

type TooManyValuesForParamError struct {
	ParamName string
	Count     int
}

func (e *TooManyValuesForParamError) Error() string {
	return fmt.Sprintf("Expected one value for %s, got %d", e.ParamName, e.Count)
}

// Handler creates http.Handler with routing matching OpenAPI spec.
func Handler(si ServerInterface) http.Handler {
	return HandlerWithOptions(si, ChiServerOptions{})
}

type ChiServerOptions struct {
	BaseURL          string
	BaseRouter       chi.Router
	Middlewares      []MiddlewareFunc
	ErrorHandlerFunc func(w http.ResponseWriter, r *http.Request, err error)
}

// HandlerFromMux creates http.Handler with routing matching OpenAPI spec based on the provided mux.
func HandlerFromMux(si ServerInterface, r chi.Router) http.Handler {
	return HandlerWithOptions(si, ChiServerOptions{
		BaseRouter: r,
	})
}

func HandlerFromMuxWithBaseURL(si ServerInterface, r chi.Router, baseURL string) http.Handler {
	return HandlerWithOptions(si, ChiServerOptions{
		BaseURL:    baseURL,
		BaseRouter: r,
	})
}

// HandlerWithOptions creates http.Handler with additional options
func HandlerWithOptions(si ServerInterface, options ChiServerOptions) http.Handler {
	r := options.BaseRouter

	if r == nil {
		r = chi.NewRouter()
	}
	if options.ErrorHandlerFunc == nil {
		options.ErrorHandlerFunc = func(w http.ResponseWriter, r *http.Request, err error) {
			http.Error(w, err.Error(), http.StatusBadRequest)
		}
	}
	wrapper := ServerInterfaceWrapper{
		Handler:            si,
		HandlerMiddlewares: options.Middlewares,
		ErrorHandlerFunc:   options.ErrorHandlerFunc,
	}

	r.Group(func(r chi.Router) {
		r.Get(options.BaseURL+"/tasks/{id}", wrapper.GetTaskByID)
	})
	r.Group(func(r chi.Router) {
		r.Patch(options.BaseURL+"/tasks/{id}", wrapper.PatchTaskByID)
	})

	return r
}