# strict-server/pkg/openapi/api/v1/api.gen.go
// Package api_v1 provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/oapi-codegen/oapi-codegen/v2 version v2.6.0 DO NOT EDIT.
package api_v1

import (
	"context"
	"encoding/json"
	"fmt"
	"net/http"
	"time"

	"github.com/fedotovmax/spec-first-openapi/transport/http/request"
	"github.com/go-chi/chi/v5"
	"github.com/oapi-codegen/runtime"
	strictnethttp "github.com/oapi-codegen/runtime/strictmiddleware/nethttp"
)

// Defines values for Status.
const (
	Backlog    Status = "backlog"
	Canceled   Status = "canceled"
	Completed  Status = "completed"
	InProgress Status = "in_progress"
	OnHold     Status = "on_hold"
)

// Valid indicates whether the value is a known member of the Status enum.
func (e Status) Valid() bool {
	switch e {
	case Backlog:
		return true
	case Canceled:
		return true
	case Completed:
		return true
	case InProgress:
		return true
	case OnHold:
		return true
	default:
		return false
	}
}

// Error defines model for error.
type Error struct {
	// Message Описание ошибки
	Message string `json:"message"`
}

// Ok defines model for ok.
type Ok struct {
	// Ok Флаг - все хорошо или нет
	Ok bool `json:"ok"`
}

// Status Текущее состояние жизненного цикла задачи: * `backlog` - Задача создана, но еще не взята в работу. * `in_progress` - Над задачей ведется работа. * `on_hold` - Работа приостановлена (ждем ответа, заблокирована). * `completed` - Задача успешно выполнена. * `canceled` - Задача отменена и не будет выполнена.
type Status string

// Task defines model for task.
type Task struct {
	// CreatedAt Время создания задачи
	CreatedAt time.Time `json:"created_at"`

	// Id ID задачи
	Id string `json:"id"`

	// IsActive Отвечает за скрытость или видимость задачи
	IsActive bool `json:"is_active"`

	// Status Текущее состояние жизненного цикла задачи: * `backlog` - Задача создана, но еще не взята в работу. * `in_progress` - Над задачей ведется работа. * `on_hold` - Работа приостановлена (ждем ответа, заблокирована). * `completed` - Задача успешно выполнена. * `canceled` - Задача отменена и не будет выполнена.
	Status Status `json:"status"`

	// Title Заголовок задачи
	Title string `json:"title"`
}

// UpdateTask defines model for update_task.
type UpdateTask struct {
	// Email Адрес эл.почти
	Email string `json:"email"`

	// Ip IP адрес
	Ip *string `json:"ip,omitempty"`

	// IsActive Отвечает за скрытость или видимость задачи
	IsActive *bool `json:"is_active,omitempty"`

	// Status Текущее состояние жизненного цикла задачи: * `backlog` - Задача создана, но еще не взята в работу. * `in_progress` - Над задачей ведется работа. * `on_hold` - Работа приостановлена (ждем ответа, заблокирована). * `completed` - Задача успешно выполнена. * `canceled` - Задача отменена и не будет выполнена.
	Status *Status `json:"status,omitempty"`

	// Title Заголовок задачи
	Title request.Nullable[string] `json:"title,omitempty"`
}

// ValidationError defines model for validation_error.
type ValidationError struct {
	// Errors Список полей с ошибками валидации
	Errors *[]ValidationErrorItem `json:"errors,omitempty"`

	// Message Описание ошибки
	Message string `json:"message"`
}

// ValidationErrorItem defines model for validation_error_item.
type ValidationErrorItem struct {
	// Error Описание ошибки
	Error string `json:"error"`

	// Field Поле у которого есть ошибка
	Field string `json:"field"`
}

// Id defines model for id.
type Id = string

// PatchTaskByIDJSONRequestBody defines body for PatchTaskByID for application/json ContentType.
type PatchTaskByIDJSONRequestBody = UpdateTask

// ServerInterface represents all server handlers.
type ServerInterface interface {
	// Get task by id
	// (GET /tasks/{id})
	GetTaskByID(w http.ResponseWriter, r *http.Request, id Id)
	// Patch task by id
	// (PATCH /tasks/{id})
	PatchTaskByID(w http.ResponseWriter, r *http.Request, id Id)
}

// Unimplemented server implementation that returns http.StatusNotImplemented for each endpoint.

type Unimplemented struct{}

// Get task by id
// (GET /tasks/{id})
func (_ Unimplemented) GetTaskByID(w http.ResponseWriter, r *http.Request, id Id) {
	w.WriteHeader(http.StatusNotImplemented)
}

// Patch task by id
// (PATCH /tasks/{id})
func (_ Unimplemented) PatchTaskByID(w http.ResponseWriter, r *http.Request, id Id) {
	w.WriteHeader(http.StatusNotImplemented)
}

// ServerInterfaceWrapper converts contexts to parameters.
type ServerInterfaceWrapper struct {
	Handler            ServerInterface
	HandlerMiddlewares []MiddlewareFunc
	ErrorHandlerFunc   func(w http.ResponseWriter, r *http.Request, err error)
}

type MiddlewareFunc func(http.Handler) http.Handler

// GetTaskByID operation middleware
func (siw *ServerInterfaceWrapper) GetTaskByID(w http.ResponseWriter, r *http.Request) {

	var err error

	// ------------- Path parameter "id" -------------
	var id Id

	err = runtime.BindStyledParameterWithOptions("simple", "id", chi.URLParam(r, "id"), &id, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true, Type: "string", Format: "uuid"})
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "id", Err: err})
		return
	}

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.GetTaskByID(w, r, id)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// PatchTaskByID operation middleware
func (siw *ServerInterfaceWrapper) PatchTaskByID(w http.ResponseWriter, r *http.Request) {

	var err error

	// ------------- Path parameter "id" -------------
	var id Id

	err = runtime.BindStyledParameterWithOptions("simple", "id", chi.URLParam(r, "id"), &id, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true, Type: "string", Format: "uuid"})
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "id", Err: err})
		return
	}

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.PatchTaskByID(w, r, id)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

type UnescapedCookieParamError struct {
	ParamName string
	Err       error
}

func (e *UnescapedCookieParamError) Error() string {
	return fmt.Sprintf("error unescaping cookie parameter '%s'", e.ParamName)
}

func (e *UnescapedCookieParamError) Unwrap() error {
	return e.Err
}

type UnmarshalingParamError struct {
	ParamName string
	Err       error
}

func (e *UnmarshalingParamError) Error() string {
	return fmt.Sprintf("Error unmarshaling parameter %s as JSON: %s", e.ParamName, e.Err.Error())
}

func (e *UnmarshalingParamError) Unwrap() error {
	return e.Err
}

type RequiredParamError struct {
	ParamName string
}

func (e *RequiredParamError) Error() string {
	return fmt.Sprintf("Query argument %s is required, but not found", e.ParamName)
}

type RequiredHeaderError struct {
	ParamName string
	Err       error
}

func (e *RequiredHeaderError) Error() string {
	return fmt.Sprintf("Header parameter %s is required, but not found", e.ParamName)
}

func (e *RequiredHeaderError) Unwrap() error {
	return e.Err
}

type InvalidParamFormatError struct {
	ParamName string
	Err       error
}

func (e *InvalidParamFormatError) Error() string {
	return fmt.Sprintf("Invalid format for parameter %s: %s", e.ParamName, e.Err.Error())
}

func (e *InvalidParamFormatError) Unwrap() error {
	return e.Err
}

type TooManyValuesForParamError struct {
	ParamName string
	Count     int
}

func (e *TooManyValuesForParamError) Error() string {
	return fmt.Sprintf("Expected one value for %s, got %d", e.ParamName, e.Count)
}

// Handler creates http.Handler with routing matching OpenAPI spec.
func Handler(si ServerInterface) http.Handler {
	return HandlerWithOptions(si, ChiServerOptions{})
}

type ChiServerOptions struct {
	BaseURL          string
	BaseRouter       chi.Router
	Middlewares      []MiddlewareFunc
	ErrorHandlerFunc func(w http.ResponseWriter, r *http.Request, err error)
}

// HandlerFromMux creates http.Handler with routing matching OpenAPI spec based on the provided mux.
func HandlerFromMux(si ServerInterface, r chi.Router) http.Handler {
	return HandlerWithOptions(si, ChiServerOptions{
		BaseRouter: r,
	})
}

func HandlerFromMuxWithBaseURL(si ServerInterface, r chi.Router, baseURL string) http.Handler {
	return HandlerWithOptions(si, ChiServerOptions{
		BaseURL:    baseURL,
		BaseRouter: r,
	})
}

// HandlerWithOptions creates http.Handler with additional options
func HandlerWithOptions(si ServerInterface, options ChiServerOptions) http.Handler {
	r := options.BaseRouter

	if r == nil {
		r = chi.NewRouter()
	}
	if options.ErrorHandlerFunc == nil {
		options.ErrorHandlerFunc = func(w http.ResponseWriter, r *http.Request, err error) {
			http.Error(w, err.Error(), http.StatusBadRequest)
		}
	}
	wrapper := ServerInterfaceWrapper{
		Handler:            si,
		HandlerMiddlewares: options.Middlewares,
		ErrorHandlerFunc:   options.ErrorHandlerFunc,
	}

	r.Group(func(r chi.Router) {
		r.Get(options.BaseURL+"/tasks/{id}", wrapper.GetTaskByID)
	})
	r.Group(func(r chi.Router) {
		r.Patch(options.BaseURL+"/tasks/{id}", wrapper.PatchTaskByID)
	})

	return r
}

type GetTaskByIDRequestObject struct {
	Id Id `json:"id"`
}

type GetTaskByIDResponseObject interface {
	VisitGetTaskByIDResponse(w http.ResponseWriter) error
}

type GetTaskByID200JSONResponse Task

func (response GetTaskByID200JSONResponse) VisitGetTaskByIDResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(200)

	return json.NewEncoder(w).Encode(response)
}

type GetTaskByID400JSONResponse ValidationError

func (response GetTaskByID400JSONResponse) VisitGetTaskByIDResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(400)

	return json.NewEncoder(w).Encode(response)
}

type GetTaskByID404JSONResponse Error

func (response GetTaskByID404JSONResponse) VisitGetTaskByIDResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(404)

	return json.NewEncoder(w).Encode(response)
}

type GetTaskByID500JSONResponse Error

func (response GetTaskByID500JSONResponse) VisitGetTaskByIDResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(500)

	return json.NewEncoder(w).Encode(response)
}

type GetTaskByIDdefaultJSONResponse struct {
	Body       Error
	StatusCode int
}

func (response GetTaskByIDdefaultJSONResponse) VisitGetTaskByIDResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(response.StatusCode)

	return json.NewEncoder(w).Encode(response.Body)
}

type PatchTaskByIDRequestObject struct {
	Id   Id `json:"id"`
	Body *PatchTaskByIDJSONRequestBody
}

type PatchTaskByIDResponseObject interface {
	VisitPatchTaskByIDResponse(w http.ResponseWriter) error
}

type PatchTaskByID200JSONResponse Ok

func (response PatchTaskByID200JSONResponse) VisitPatchTaskByIDResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(200)

	return json.NewEncoder(w).Encode(response)
}

type PatchTaskByID400JSONResponse ValidationError

func (response PatchTaskByID400JSONResponse) VisitPatchTaskByIDResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(400)

	return json.NewEncoder(w).Encode(response)
}

type PatchTaskByID404JSONResponse Error

func (response PatchTaskByID404JSONResponse) VisitPatchTaskByIDResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(404)

	return json.NewEncoder(w).Encode(response)
}

type PatchTaskByID500JSONResponse Error

func (response PatchTaskByID500JSONResponse) VisitPatchTaskByIDResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(500)

	return json.NewEncoder(w).Encode(response)
}

type PatchTaskByIDdefaultJSONResponse struct {
	Body       Error
	StatusCode int
}

func (response PatchTaskByIDdefaultJSONResponse) VisitPatchTaskByIDResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(response.StatusCode)

	return json.NewEncoder(w).Encode(response.Body)
}

// StrictServerInterface represents all server handlers.
type StrictServerInterface interface {
	// Get task by id
	// (GET /tasks/{id})
	GetTaskByID(ctx context.Context, request GetTaskByIDRequestObject) (GetTaskByIDResponseObject, error)
	// Patch task by id
	// (PATCH /tasks/{id})
	PatchTaskByID(ctx context.Context, request PatchTaskByIDRequestObject) (PatchTaskByIDResponseObject, error)
}

type StrictHandlerFunc = strictnethttp.StrictHTTPHandlerFunc
type StrictMiddlewareFunc = strictnethttp.StrictHTTPMiddlewareFunc

type StrictHTTPServerOptions struct {
	RequestErrorHandlerFunc  func(w http.ResponseWriter, r *http.Request, err error)
	ResponseErrorHandlerFunc func(w http.ResponseWriter, r *http.Request, err error)
}

func NewStrictHandler(ssi StrictServerInterface, middlewares []StrictMiddlewareFunc) ServerInterface {
	return &strictHandler{ssi: ssi, middlewares: middlewares, options: StrictHTTPServerOptions{
		RequestErrorHandlerFunc: func(w http.ResponseWriter, r *http.Request, err error) {
			http.Error(w, err.Error(), http.StatusBadRequest)
		},
		ResponseErrorHandlerFunc: func(w http.ResponseWriter, r *http.Request, err error) {
			http.Error(w, err.Error(), http.StatusInternalServerError)
		},
	}}
}

func NewStrictHandlerWithOptions(ssi StrictServerInterface, middlewares []StrictMiddlewareFunc, options StrictHTTPServerOptions) ServerInterface {
	return &strictHandler{ssi: ssi, middlewares: middlewares, options: options}
}

type strictHandler struct {
	ssi         StrictServerInterface
	middlewares []StrictMiddlewareFunc
	options     StrictHTTPServerOptions
}

// GetTaskByID operation middleware
func (sh *strictHandler) GetTaskByID(w http.ResponseWriter, r *http.Request, id Id) {
	var request GetTaskByIDRequestObject

	request.Id = id

	handler := func(ctx context.Context, w http.ResponseWriter, r *http.Request, request interface{}) (interface{}, error) {
		return sh.ssi.GetTaskByID(ctx, request.(GetTaskByIDRequestObject))
	}
	for _, middleware := range sh.middlewares {
		handler = middleware(handler, "GetTaskByID")
	}

	response, err := handler(r.Context(), w, r, request)

	if err != nil {
		sh.options.ResponseErrorHandlerFunc(w, r, err)
	} else if validResponse, ok := response.(GetTaskByIDResponseObject); ok {
		if err := validResponse.VisitGetTaskByIDResponse(w); err != nil {
			sh.options.ResponseErrorHandlerFunc(w, r, err)
		}
	} else if response != nil {
		sh.options.ResponseErrorHandlerFunc(w, r, fmt.Errorf("unexpected response type: %T", response))
	}
}

// PatchTaskByID operation middleware
func (sh *strictHandler) PatchTaskByID(w http.ResponseWriter, r *http.Request, id Id) {
	var request PatchTaskByIDRequestObject

	request.Id = id

	var body PatchTaskByIDJSONRequestBody
	if err := json.NewDecoder(r.Body).Decode(&body); err != nil {
		sh.options.RequestErrorHandlerFunc(w, r, fmt.Errorf("can't decode JSON body: %w", err))
		return
	}
	request.Body = &body

	handler := func(ctx context.Context, w http.ResponseWriter, r *http.Request, request interface{}) (interface{}, error) {
		return sh.ssi.PatchTaskByID(ctx, request.(PatchTaskByIDRequestObject))
	}
	for _, middleware := range sh.middlewares {
		handler = middleware(handler, "PatchTaskByID")
	}

	response, err := handler(r.Context(), w, r, request)

	if err != nil {
		sh.options.ResponseErrorHandlerFunc(w, r, err)
	} else if validResponse, ok := response.(PatchTaskByIDResponseObject); ok {
		if err := validResponse.VisitPatchTaskByIDResponse(w); err != nil {
			sh.options.ResponseErrorHandlerFunc(w, r, err)
		}
	} else if response != nil {
		sh.options.ResponseErrorHandlerFunc(w, r, fmt.Errorf("unexpected response type: %T", response))
	}
}